C++ code uploaded
animat.h: Class for the simulated creatures, called Animats. Can not remember where I picked up this name, but it is unfortunately not my own.

animatsMain.cpp: Runs the simulation, piecing together the classes.

DeathFunctionTests.cpp: Written to test for a better algorithm for determining the animats that die in a generation.

generation.h: Keeps track of generational statistics.

manager.h: Used for setting up simulations and specifying end conditions.